+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - don't cache zones forever (for example removed permissions should lead to removed targets in the platform)
 - include tags in the discovery of Lambda functions (requires new permission `tag:GetResources`)
 - add vpc name to targets (requires new permission `ec2:DescribeVpcs`, can be disabled by `STEADYBIT_EXTENSION_DISCOVERY_DISABLED_VPC`)
+- add subnet target discovery
 
 ## v2.3.6